strife: Turns out that 1.0 was actually 1.1 all along.
The documentation and evidence for this is sparse.  version.txt, for
example, still shows "STRIFE(TM) VERSION 1.0", and the readme.txt
reflects this.

What settles it, however, is that the demo format writes out 0x65 into
the first byte of the format, decimal value of 101 signifying that the
engine is really 1.1 at this point.  There is an older Strife 1.0
released as a teaser demo only.  See /roguestuff/strife10.zip at your
friendly neighborhood idgames mirror, for that version specifically.